# healthplannerPythonAPI-master/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "description": "REST API for Health Advisor Application",
    "version": "1.0.0",
    "title": "Health Advisor API",
    "license": {
      "name": "IBM",
      "url": "http://www.ibm.com"
    }
  },
  "paths": {
    "/request": {
      "get": {
        "tags": [
          "Patient Queries"
        ],
        "summary": "Returns all patients",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/patient"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patient Queries"
        ],
        "summary": "Creates a new patient",
        "requestBody": {
          "description": "Patient Details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/patient"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          },
          "404": {
            "description": "Failed. Patient already exists."
          }
        }
      }
    },
    "/request/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the patient that we want to match",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Patient Queries"
        ],
        "summary": "Rerieves patient record, provided a valid ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Patient ID",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/patient"
            }
          },
          "400": {
            "description": "Failed. Misunderstood Request."
          },
          "404": {
            "description": "Failed. Patient not found."
          }
        }
      },
      "put": {
        "summary": "Edits a patient record, provided a valid ID and data",
        "tags": [
          "Patient Queries"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Patient ID",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "requestBody": {
          "description": "Patient Request Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/patient"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/patient"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      },
      "delete": {
        "summary": "Deletea a patient record, provided a valid ID",
        "tags": [
          "Patient Queries"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Patient ID",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          },
          "400": {
            "description": "Failed. Misunderstood Request."
          },
          "404": {
            "description": "Failed. Patient Request not found."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "id": {
        "properties": {
          "uuid": {
            "type": "string"
          }
        }
      },
      "patient": {
        "type": "object",
        "properties": {
            "id": {
              "type": "string",
              "format": "string"
            },
            "patientName": {
              "$ref": "#/components/schemas/patientname"
            },
            "gender": {
              "type": "string",
              "format": "string"
            },
            "dateOfBirth": {
              "type": "string",
              "format": "date-time"
            },
            "mailId": {
              "type": "string",
              "format": "email"
            },
            "maritalStatus": {
              "type": "string",
              "format": "string"
            },
            "medHistory": {
              "$ref": "#/components/schemas/medHistory"
            },
            "phone": {
              "type": "string",
              "format": "string"
            },
            "postalAddress": {
              "$ref": "#/components/schemas/postalAddress"
            }
          }
        },
        "patientname": {
          "type": "object",
          "properties": {
              "firstName": {
                "type": "string",
                "format": "string"
              },
              "lastName": {
                "type": "string",
                "format": "string"
              }
            }
        },
        "medHistory": {
          "type": "array",
          "items": {
              "type": "string"
            }
        },
        "postalAddress": {
          "type": "object",
          "properties": {
              "city": {
                "type": "string",
                "format": "string"
              },
              "line1": {
                "type": "string",
                "format": "string"
              },
              "line2": {
                "type": "string",
                "format": "string"
              },
              "state": {
                "type": "string",
                "format": "string"
              },
              "zip": {
                "type": "string",
                "format": "string"
              }
          }
        }
    }
  }
}